Conflicts:
	examples/audiorecorder/audiorecorder.cpp
	examples/audiorecorder/audiorecorder.h
	examples/cameracapture/cameracapture.cpp
	examples/cameracapture/cameracapture.h
	examples/examples.pro
	examples/player/player.cpp
	examples/player/playlistmodel.cpp
	examples/streamplayer/main.cpp
	examples/streamplayer/streamplayer.cpp
	examples/streamplayer/streamplayer.h
	examples/streamplayer/streamplayer.pro
	multimedia/qabstractmediaobject.cpp
	multimedia/qabstractmediaobject_p.h
	multimedia/qabstractmediaservice.cpp
	multimedia/qaudioinputdevicecontrol.cpp
	multimedia/qaudiorecorderservice.cpp
	multimedia/qaudiorecorderservice.h
	multimedia/qaudiosource.h
	multimedia/qcamera.cpp
	multimedia/qcameraservice.cpp
	multimedia/qcameraservice.h
	multimedia/qmediaimageviewerservice.cpp
	multimedia/qmediaimageviewerservice_p.h
	multimedia/qmediaplayer.cpp
	multimedia/qmediaplayerservice.cpp
	multimedia/qmediaplayerservice.h
	multimedia/qmediaplaylist.cpp
	multimedia/qmediarecorder.cpp
	multimedia/qmediarecorderservice.cpp
	multimedia/qmediaserviceprovider.cpp
	multimedia/qmediaslideshow.cpp
	multimedia/qmetadataprovidercontrol.cpp
	multimedia/qradioplayer.cpp
	plugins/audiocapture/audiocaptureservice.cpp
	plugins/audiocapture/audiocaptureserviceplugin.cpp
	plugins/audiocapture/audiocapturesession.cpp
	plugins/audiocapture/audiodevicecontrol.cpp
	plugins/audiocapture/audioencodecontrol.cpp
	plugins/audiocapture/audioencodecontrol.h
	plugins/audiocapture/audioinputdevicecontrol.cpp
	plugins/gstreamer/mediacapture/qgstreamercaptureservice.cpp
	plugins/gstreamer/mediacapture/qgstreamercaptureservice.h
	plugins/gstreamer/mediacapture/qgstreamercapturesession.cpp
	plugins/gstreamer/mediaplayer/qgstreamerplayercontrol.cpp
	plugins/gstreamer/mediaplayer/qgstreamerplayerservice.h
	plugins/gstreamer/mediaplayer/qgstreamerplayersession.cpp
	plugins/gstreamer/mediaplayer/qgstreamerplayersession.h
	plugins/gstreamer/qgstreamervideoinputdevicecontrol.h
	plugins/phonon/qphononplayerservice.h
	plugins/v4l/v4lserviceplugin.cpp
	plugins/wmp/qwmpplayerservice.h
	tests/auto/qabstractmediaobject/tst_qabstractmediaobject.cpp
	tests/auto/qcamera/tst_qcamera.cpp
	tests/auto/qmediaplayer/tst_qmediaplayer.cpp
	tests/auto/qmediarecorder/tst_qmediarecorder.cpp
	tests/auto/qradioplayer/tst_qradioplayer.cpp

+++ multimedia/qvideoencodercontrol.cpp
@@ -157,7 +157,7 @@ QVideoEncoderControl::~QVideoEncoderControl()
 */
 
 /*!
-  \fn QVideoEncoderControl::setQuality(qreal value)
+  \fn QVideoEncoderControl::setQuality(int value)
 
   Sets the video quality to \a value.
